reorganize new images carousel
